moved function to utils.py
