+++ bibcat/llm/evaluate.py
@@ -4,11 +4,11 @@ import numpy as np
 import pandas as pd
 
 from bibcat import config
-from bibcat.core import fetch_keyword_object
 from bibcat.core import parameters as params
 from bibcat.core.operator import Operator
 from bibcat.core.paper import Paper
 from bibcat.llm.io import get_source, read_output, write_summary
+from bibcat.utils import fetch_keyword_object
 from bibcat.utils.logger_config import setup_logger
 
 # set up global operator
@@ -412,7 +412,7 @@ def identify_missions_in_text(missions: list, text: str) -> list:
 
         # get the relevant mission keyword
         try:
-            keyword = fetch_keyword_object.fetch_keyword_object(op.keyword_objs, mission, verbose=op.verbose)
+            keyword = fetch_keyword_object(op.keyword_objs, mission, verbose=op.verbose)
         except ValueError:
             # if the keyword doesn't exist, just use the provided mission name
             keywd = mission
