+++ bibcat/tests/core/test_operator.py
@@ -8,9 +8,9 @@ import unittest
 
 from test_config import test_dict_lookup_kobj
 
-from bibcat.core import fetch_keyword_object
 from bibcat.core import parameters as params
 from bibcat.core.operator import Operator
+from bibcat.utils import fetch_keyword_object
 
 
 # Purpose: Testing the Operator class
@@ -36,7 +36,7 @@ class TestOperator(unittest.TestCase):
             for key1 in dict_acts:
                 # Otherwise, check generated modif
                 curr_lookup = key1
-                test_res = fetch_keyword_object.fetch_keyword_object(
+                test_res = fetch_keyword_object(
                     tmp_kobj_list, lookup=curr_lookup, do_raise_emptyerror=True, verbose=False
                 )
                 curr_answer = test_dict_lookup_kobj[dict_acts[key1]]
